+++ CHANGES.md
@@ -1,5 +1,9 @@
 # CHANGES
 
+## Rib-Heroku 0.8.2 -- 2012-07-20
+
+* Updated the monkey patch from latest heroku gem
+
 ## Rib-Heroku 0.8.1 -- 2012-07-19
 
 * Fixed requiring order for latest heroku